Part of the fix for BR #42177 (Normal): Kspread: Format Cell Shrinks Cell Size
We need to use double instead of int for height/width when we use koUnit for the display/input

svn path=/trunk/koffice/; revision=154140

+++ kspread/kspread_canvas.h
@@ -501,7 +501,7 @@ public:
     KSpreadHBorder( QWidget *_parent, KSpreadCanvas *_canvas, KSpreadView *_view  );
 
     int markerColumn() { return  m_iSelectionAnchor; }
-    void resizeColumn( int resize, int nb = -1, bool makeUndo=true );
+    void resizeColumn( double resize, int nb = -1, bool makeUndo=true );
     void adjustColumn( int _col = -1, bool makeUndo=true );
     void equalizeColumn( int resize );
 
@@ -567,7 +567,7 @@ public:
     KSpreadVBorder( QWidget *_parent, KSpreadCanvas *_canvas, KSpreadView *_view );
 
     int markerRow() { return  m_iSelectionAnchor; }
-    void resizeRow( int resize, int nb = -1, bool makeUndo=true );
+    void resizeRow( double resize, int nb = -1, bool makeUndo=true );
     void adjustRow( int _row = -1, bool makeUndo=true );
     void equalizeRow( int resize );
